added recommendations feature + setup for search
The recommendations system is just the search function with an empty parameter.
All that needs to be done is to pass a query to that function and the user will see search results.

The plan is to add a searchbox and improve the recommendations system by having suggested queries/ tags that the user can click on.
Also, the url scheme will be changed to reflect search results vs video vs user (a channel).