Fixed a sneaky and previously undetected syntax error
Slowly but surely, it threatened to destroy everything.
Sure, it didn't cause any noticable performance problems or linting errors but, when I tried to edit the code, it clearly was not right.
There was a forEach inside of a promise, inside of an else statemen, inside an xhr, inside a function (or at least that's how it was supposed to be).
Took a lot of work to figure out the exact error and fix it.